Support SSL and non-SSL urls, increase incoming buffer size.
We'll just blindly assume that if the connection to the server fails, that it was because it was a 400 error for SSL not being supported, in which case we'll just fall back to using a non-SSL connection (aka wss:// vs ws://)

Also, finally found the call for increasing the client's buffer size! Increased the inbound data buffer's size from 64 KB to 256 KB, which should hopefully be enough.