Add archipelago_rs@godot_client binaries and GDNS wrappers
These point to my fork of archipelago_rs (SpenserHaddad/archipelago_rs, branch godot_client) with some test work on a GDNative client.

I originally wrote this because I thought that Godot 3.5's built-in WebSocket support was insufficient for an Archipelago client. I have since been proven wrong, so this is being left in a not-quite-working version.